+++ CHANGELOG.md
@@ -5,6 +5,7 @@ pandoc-plot uses [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 Release 0.6.0.0
 --------------
 
+* The determination of which figures to re-render or not has been improved. For example, changing the caption will not trigger a re-render of a figure anymore.
 * Added the ability to write the example configuration to an arbitrary file using `pandoc-plot write-example-config`.
 * Added the possibility to specify the configuration file via metadata. For example, in Markdown: